renamed pages.js
